Move processor editor parameter classes into a separate module.
