+++ kunquat/tracker/ui/model/processor.py
@@ -12,33 +12,33 @@
 #
 
 from kunquat.kunquat.kunquat import get_default_value
-from procparamsadd import ProcParamsAdd
-from procparamsdelay import ProcParamsDelay
-from procparamsenvgen import ProcParamsEnvgen
-from procparamsfilter import ProcParamsFilter
-from procparamsforce import ProcParamsForce
-from procparamsfreeverb import ProcParamsFreeverb
-from procparamsgaincomp import ProcParamsGainComp
-from procparamspanning import ProcParamsPanning
-from procparamspitch import ProcParamsPitch
-from procparamsringmod import ProcParamsRingmod
-from procparams_stream import ProcParamsStream
-from procparamsvolume import ProcParamsVolume
+from procparams.addparams import AddParams
+from procparams.delayparams import DelayParams
+from procparams.envgenparams import EnvgenParams
+from procparams.filterparams import FilterParams
+from procparams.forceparams import ForceParams
+from procparams.freeverbparams import FreeverbParams
+from procparams.gaincompparams import GainCompParams
+from procparams.panningparams import PanningParams
+from procparams.pitchparams import PitchParams
+from procparams.ringmodparams import RingmodParams
+from procparams.streamparams import StreamParams
+from procparams.volumeparams import VolumeParams
 
 
 _proc_classes = {
-    'add':      ProcParamsAdd,
-    'delay':    ProcParamsDelay,
-    'envgen':   ProcParamsEnvgen,
-    'filter':   ProcParamsFilter,
-    'force':    ProcParamsForce,
-    'freeverb': ProcParamsFreeverb,
-    'gaincomp': ProcParamsGainComp,
-    'panning':  ProcParamsPanning,
-    'pitch':    ProcParamsPitch,
-    'ringmod':  ProcParamsRingmod,
-    'stream':   ProcParamsStream,
-    'volume':   ProcParamsVolume,
+    'add':      AddParams,
+    'delay':    DelayParams,
+    'envgen':   EnvgenParams,
+    'filter':   FilterParams,
+    'force':    ForceParams,
+    'freeverb': FreeverbParams,
+    'gaincomp': GainCompParams,
+    'panning':  PanningParams,
+    'pitch':    PitchParams,
+    'ringmod':  RingmodParams,
+    'stream':   StreamParams,
+    'volume':   VolumeParams,
 }
