+++ kunquat/tracker/ui/model/procparams/freeverbparams.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 #
-# Author: Tomi Jylhä-Ollila, Finland 2016
+# Author: Tomi Jylhä-Ollila, Finland 2015-2016
 #
 # This file is part of Kunquat.
 #
@@ -14,7 +14,7 @@
 from procparams import ProcParams
 
 
-class ProcParamsPanning(ProcParams):
+class FreeverbParams(ProcParams):
 
     @staticmethod
     def get_default_signal_type():
@@ -23,9 +23,8 @@ class ProcParamsPanning(ProcParams):
     @staticmethod
     def get_port_info():
         return {
-            'in_00':  u'pan',
-            'in_01':  u'audio L',
-            'in_02':  u'audio R',
+            'in_00':  u'audio L',
+            'in_01':  u'audio R',
             'out_00': u'audio L',
             'out_01': u'audio R',
         }
@@ -33,10 +32,16 @@ class ProcParamsPanning(ProcParams):
     def __init__(self, proc_id, controller):
         ProcParams.__init__(self, proc_id, controller)
 
-    def get_panning(self):
-        return self._get_value('p_f_panning.json', 0.0)
+    def get_reflectivity(self):
+        return self._get_value('p_f_refl.json', 20)
 
-    def set_panning(self, value):
-        self._set_value('p_f_panning.json', value)
+    def set_reflectivity(self, value):
+        self._set_value('p_f_refl.json', value)
+
+    def get_damp(self):
+        return self._get_value('p_f_damp.json', 20)
+
+    def set_damp(self, value):
+        self._set_value('p_f_damp.json', value)
