+++ kunquat/tracker/ui/model/procparams/gaincompparams.py
@@ -14,7 +14,7 @@
 from procparams import ProcParams
 
 
-class ProcParamsFilter(ProcParams):
+class GainCompParams(ProcParams):
 
     @staticmethod
     def get_default_signal_type():
@@ -23,10 +23,8 @@ class ProcParamsFilter(ProcParams):
     @staticmethod
     def get_port_info():
         return {
-            'in_00':  u'cutoff',
-            'in_01':  u'reso',
-            'in_02':  u'audio L',
-            'in_03':  u'audio R',
+            'in_00':  u'audio L',
+            'in_01':  u'audio R',
             'out_00': u'audio L',
             'out_01': u'audio R',
         }
@@ -34,16 +32,19 @@ class ProcParamsFilter(ProcParams):
     def __init__(self, proc_id, controller):
         ProcParams.__init__(self, proc_id, controller)
 
-    def get_cutoff(self):
-        return self._get_value('p_f_cutoff.json', 100.0)
+    def get_mapping_enabled(self):
+        return self._get_value('p_b_map_enabled.json', False)
 
-    def set_cutoff(self, value):
-        self._set_value('p_f_cutoff.json', value)
+    def set_mapping_enabled(self, enabled):
+        self._set_value('p_b_map_enabled.json', enabled)
 
-    def get_resonance(self):
-        return self._get_value('p_f_resonance.json', 0.0)
+    def get_mapping(self):
+        ret_env = { 'nodes': [ [0, 0], [1, 1] ], 'smooth': False }
+        stored_env = self._get_value('p_e_map.json', None) or {}
+        ret_env.update(stored_env)
+        return ret_env
 
-    def set_resonance(self, value):
-        self._set_value('p_f_resonance.json', value)
+    def set_mapping(self, envelope):
+        self._set_value('p_e_map.json', envelope)
