+++ kunquat/tracker/ui/model/procparams/panningparams.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 #
-# Author: Tomi Jylhä-Ollila, Finland 2015-2016
+# Author: Tomi Jylhä-Ollila, Finland 2016
 #
 # This file is part of Kunquat.
 #
@@ -14,7 +14,7 @@
 from procparams import ProcParams
 
 
-class ProcParamsVolume(ProcParams):
+class PanningParams(ProcParams):
 
     @staticmethod
     def get_default_signal_type():
@@ -23,7 +23,7 @@ class ProcParamsVolume(ProcParams):
     @staticmethod
     def get_port_info():
         return {
-            'in_00':  u'force',
+            'in_00':  u'pan',
             'in_01':  u'audio L',
             'in_02':  u'audio R',
             'out_00': u'audio L',
@@ -33,10 +33,10 @@ class ProcParamsVolume(ProcParams):
     def __init__(self, proc_id, controller):
         ProcParams.__init__(self, proc_id, controller)
 
-    def get_volume(self):
-        return self._get_value('p_f_volume.json', 0.0)
+    def get_panning(self):
+        return self._get_value('p_f_panning.json', 0.0)
 
-    def set_volume(self, value):
-        self._set_value('p_f_volume.json', value)
+    def set_panning(self, value):
+        self._set_value('p_f_panning.json', value)
